fixes #9
Added a function that trims the read sequence and associated quality
line given the start and end position to keep.
This function will be called for both mate for each read. Therefore,
giving the function the Python 0-based coordinates to keep avoids
re-computing them every time the function is called. Smart huh?